requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg[binary]
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
python-multipart
pydantic
pydantic-settings
typing_extensions
python-dotenv
requests
aiofiles
apscheduler
praytimes
hijri-converter
timezonefinder
httpx
aiocache
pydub
plyer
pytz
pygame
pydantic[email]
aiosmtplib
jinja2
resend
boto3